# api/sys/v1/sys_user_router.pb.go
// Code generated by https://github.com/zhufuyi/sponge, DO NOT EDIT.

package v1

import (
	context "context"
	gin "github.com/gin-gonic/gin"
	errcode "github.com/zhufuyi/sponge/pkg/errcode"
	middleware "github.com/zhufuyi/sponge/pkg/gin/middleware"
	zap "go.uber.org/zap"
	strings "strings"
)

// import packages: strings. context. errcode. middleware. zap. gin.

type SysUserLogicer interface {
	Create(ctx context.Context, req *CreateSysUserRequest) (*CreateSysUserReply, error)
	DeleteByID(ctx context.Context, req *DeleteSysUserByIDRequest) (*DeleteSysUserByIDReply, error)
	DeleteByIDs(ctx context.Context, req *DeleteSysUserByIDsRequest) (*DeleteSysUserByIDsReply, error)
	UpdateByID(ctx context.Context, req *UpdateSysUserByIDRequest) (*UpdateSysUserByIDReply, error)
	GetByID(ctx context.Context, req *GetSysUserByIDRequest) (*GetSysUserByIDReply, error)
	GetByCondition(ctx context.Context, req *GetSysUserByConditionRequest) (*GetSysUserByConditionReply, error)
	ListByIDs(ctx context.Context, req *ListSysUserByIDsRequest) (*ListSysUserByIDsReply, error)
	ListByLastID(ctx context.Context, req *ListSysUserByLastIDRequest) (*ListSysUserByLastIDReply, error)
	List(ctx context.Context, req *ListSysUserRequest) (*ListSysUserReply, error)
}

type SysUserOption func(*sysUserOptions)

type sysUserOptions struct {
	isFromRPC  bool
	responser  errcode.Responser
	zapLog     *zap.Logger
	httpErrors []*errcode.Error
	rpcStatus  []*errcode.RPCStatus
	wrapCtxFn  func(c *gin.Context) context.Context
}

func (o *sysUserOptions) apply(opts ...SysUserOption) {
	for _, opt := range opts {
		opt(o)
	}
}

func WithSysUserHTTPResponse() SysUserOption {
	return func(o *sysUserOptions) {
		o.isFromRPC = false
	}
}

func WithSysUserRPCResponse() SysUserOption {
	return func(o *sysUserOptions) {
		o.isFromRPC = true
	}
}

func WithSysUserResponser(responser errcode.Responser) SysUserOption {
	return func(o *sysUserOptions) {
		o.responser = responser
	}
}

func WithSysUserLogger(zapLog *zap.Logger) SysUserOption {
	return func(o *sysUserOptions) {
		o.zapLog = zapLog
	}
}

func WithSysUserErrorToHTTPCode(e ...*errcode.Error) SysUserOption {
	return func(o *sysUserOptions) {
		o.httpErrors = e
	}
}

func WithSysUserRPCStatusToHTTPCode(s ...*errcode.RPCStatus) SysUserOption {
	return func(o *sysUserOptions) {
		o.rpcStatus = s
	}
}

func WithSysUserWrapCtx(wrapCtxFn func(c *gin.Context) context.Context) SysUserOption {
	return func(o *sysUserOptions) {
		o.wrapCtxFn = wrapCtxFn
	}
}

func RegisterSysUserRouter(
	iRouter gin.IRouter,
	groupPathMiddlewares map[string][]gin.HandlerFunc,
	singlePathMiddlewares map[string][]gin.HandlerFunc,
	iLogic SysUserLogicer,
	opts ...SysUserOption) {

	o := &sysUserOptions{}
	o.apply(opts...)

	if o.responser == nil {
		o.responser = errcode.NewResponser(o.isFromRPC, o.httpErrors, o.rpcStatus)
	}
	if o.zapLog == nil {
		o.zapLog, _ = zap.NewProduction()
	}

	r := &sysUserRouter{
		iRouter:               iRouter,
		groupPathMiddlewares:  groupPathMiddlewares,
		singlePathMiddlewares: singlePathMiddlewares,
		iLogic:                iLogic,
		iResponse:             o.responser,
		zapLog:                o.zapLog,
		wrapCtxFn:             o.wrapCtxFn,
	}
	r.register()
}

type sysUserRouter struct {
	iRouter               gin.IRouter
	groupPathMiddlewares  map[string][]gin.HandlerFunc
	singlePathMiddlewares map[string][]gin.HandlerFunc
	iLogic                SysUserLogicer
	iResponse             errcode.Responser
	zapLog                *zap.Logger
	wrapCtxFn             func(c *gin.Context) context.Context
}

func (r *sysUserRouter) register() {
	r.iRouter.Handle("POST", "/v1/sys_user/create", r.withMiddleware("POST", "/v1/sys_user/create", r.Create_2)...)
	r.iRouter.Handle("DELETE", "/v1/sys_user/delete/:id", r.withMiddleware("DELETE", "/v1/sys_user/delete/:id", r.DeleteByID_2)...)
	r.iRouter.Handle("POST", "/v1/sys_user/delete/batch", r.withMiddleware("POST", "/v1/sys_user/delete/batch", r.DeleteByIDs_2)...)
	r.iRouter.Handle("PUT", "/v1/sys_user/update/:id", r.withMiddleware("PUT", "/v1/sys_user/update/:id", r.UpdateByID_2)...)
	r.iRouter.Handle("GET", "/v1/sys_user/get/:id", r.withMiddleware("GET", "/v1/sys_user/get/:id", r.GetByID_2)...)
	r.iRouter.Handle("POST", "/v1/sys_user/getByCondition", r.withMiddleware("POST", "/v1/sys_user/getByCondition", r.GetByCondition_2)...)
	r.iRouter.Handle("POST", "/v1/sys_user/listByIDs", r.withMiddleware("POST", "/v1/sys_user/listByIDs", r.ListByIDs_2)...)
	r.iRouter.Handle("POST", "/v1/sys_user/listByLastID", r.withMiddleware("POST", "/v1/sys_user/listByLastID", r.ListByLastID_2)...)
	r.iRouter.Handle("POST", "/v1/sys_user/list", r.withMiddleware("POST", "/v1/sys_user/list", r.List_2)...)

}

func (r *sysUserRouter) withMiddleware(method string, path string, fn gin.HandlerFunc) []gin.HandlerFunc {
	handlerFns := []gin.HandlerFunc{}

	// determine if a route group is hit or miss, left prefix rule
	for groupPath, fns := range r.groupPathMiddlewares {
		if groupPath == "" || groupPath == "/" {
			handlerFns = append(handlerFns, fns...)
			continue
		}
		size := len(groupPath)
		if len(path) < size {
			continue
		}
		if groupPath == path[:size] {
			handlerFns = append(handlerFns, fns...)
		}
	}

	// determine if a single route has been hit
	key := strings.ToUpper(method) + "->" + path
	if fns, ok := r.singlePathMiddlewares[key]; ok {
		handlerFns = append(handlerFns, fns...)
	}

	return append(handlerFns, fn)
}

var _ middleware.CtxKeyString

func (r *sysUserRouter) Create_2(c *gin.Context) {
	req := &CreateSysUserRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.Create(ctx, req)
	if err != nil {
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *sysUserRouter) DeleteByID_2(c *gin.Context) {
	req := &DeleteSysUserByIDRequest{}
	var err error

	if err = c.ShouldBindUri(req); err != nil {
		r.zapLog.Warn("ShouldBindUri error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	if err = c.ShouldBindQuery(req); err != nil {
		r.zapLog.Warn("ShouldBindQuery error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.DeleteByID(ctx, req)
	if err != nil {
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *sysUserRouter) DeleteByIDs_2(c *gin.Context) {
	req := &DeleteSysUserByIDsRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.DeleteByIDs(ctx, req)
	if err != nil {
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *sysUserRouter) UpdateByID_2(c *gin.Context) {
	req := &UpdateSysUserByIDRequest{}
	var err error

	if err = c.ShouldBindUri(req); err != nil {
		r.zapLog.Warn("ShouldBindUri error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.UpdateByID(ctx, req)
	if err != nil {
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *sysUserRouter) GetByID_2(c *gin.Context) {
	req := &GetSysUserByIDRequest{}
	var err error

	if err = c.ShouldBindUri(req); err != nil {
		r.zapLog.Warn("ShouldBindUri error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	if err = c.ShouldBindQuery(req); err != nil {
		r.zapLog.Warn("ShouldBindQuery error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.GetByID(ctx, req)
	if err != nil {
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *sysUserRouter) GetByCondition_2(c *gin.Context) {
	req := &GetSysUserByConditionRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.GetByCondition(ctx, req)
	if err != nil {
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *sysUserRouter) ListByIDs_2(c *gin.Context) {
	req := &ListSysUserByIDsRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.ListByIDs(ctx, req)
	if err != nil {
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *sysUserRouter) ListByLastID_2(c *gin.Context) {
	req := &ListSysUserByLastIDRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.ListByLastID(ctx, req)
	if err != nil {
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *sysUserRouter) List_2(c *gin.Context) {
	req := &ListSysUserRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.List(ctx, req)
	if err != nil {
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}