# docs/apis.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "sys api docs",
    "version": "2.0"
  },
...
  "host": "localhost:8080",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/sys_user/create": {
      "post": {
        "summary": "创建用户",
        "operationId": "create_sys_user",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateSysUserReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateSysUserRequest"
            }
          }
        ],
        "tags": [
          "sys_user"
        ]
      }
    },
    "/v1/sys_user/delete/batch": {
      "post": {
        "summary": "批量删除用户",
        "operationId": "delete_sys_user_by_ids",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteSysUserByIDsReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "ids",
            "in": "body",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "uint64"
              }
            }
          }
        ],
        "tags": [
          "sys_user"
        ]
      }
    },
    "/v1/sys_user/delete/{id}": {
      "delete": {
        "summary": "删除用户",
        "operationId": "delete_sys_user_by_id",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteSysUserByIDReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "uint64"
          }
        ],
        "tags": [
          "sys_user"
        ]
      }
    },
    "/v1/sys_user/get/{id}": {
      "get": {
        "summary": "通过ID获取用户",
        "operationId": "get_sys_user_by_id",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetSysUserByIDReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "uint64"
          }
        ],
        "tags": [
          "sys_user"
        ]
      }
    },
    "/v1/sys_user/getByCondition": {
      "post": {
        "summary": "通过条件获取用户",
        "operationId": "get_sys_user_by_condition",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetSysUserByConditionReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1GetSysUserByConditionRequest"
            }
          }
        ],
        "tags": [
          "sys_user"
        ]
      }
    },
    "/v1/sys_user/list": {
      "post": {
        "summary": "获取用户列表",
        "operationId": "list_sys_user",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListSysUserReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1ListSysUserRequest"
            }
          }
        ],
        "tags": [
          "sys_user"
        ]
      }
    },
    "/v1/sys_user/listByIDs": {
      "post": {
        "summary": "通过ID批量获取用户",
        "operationId": "list_sys_user_by_ids",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListSysUserByIDsReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "ids",
            "in": "body",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "uint64"
              }
            }
          }
        ],
        "tags": [
          "sys_user"
        ]
      }
    },
    "/v1/sys_user/listByLastID": {
      "post": {
        "summary": "通过最后ID获取用户列表",
        "operationId": "list_sys_user_by_last_id",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListSysUserByLastIDReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1ListSysUserByLastIDRequest"
            }
          }
        ],
        "tags": [
          "sys_user"
        ]
      }
    },
    "/v1/sys_user/update/{id}": {
      "put": {
        "summary": "通过ID更新用户",
        "operationId": "update_sys_user_by_id",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdateSysUserByIDReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "description": "用户ID",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "uint64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1sysUserUpdateByIDBody"
            }
          }
        ],
        "tags": [
          "sys_user"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "typesColumn": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "title": "column name"
        },
        "exp": {
          "type": "string",
          "title": "expressions, which default to = when the value is null, have =, !=, \u003e, \u003e=, \u003c, \u003c=, like, in"
        },
        "value": {
          "type": "string",
          "title": "column value"
        },
        "logic": {
          "type": "string",
          "title": "logical type, defaults to and when value is null, only \u0026(and), ||(or)"
        }
      }
    },
    "typesConditions": {
      "type": "object",
      "properties": {
        "columns": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/typesColumn"
          },
          "title": "query conditions"
        }
      }
    },
    "typesParams": {
      "type": "object",
      "properties": {
        "page": {
          "type": "integer",
          "format": "int32",
          "title": "page number, starting from 0"
        },
        "limit": {
          "type": "integer",
          "format": "int32",
          "title": "lines per page"
        },
        "sort": {
          "type": "string",
          "title": "sorted fields, multi-column sorting separated by commas"
        },
        "columns": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/typesColumn"
          },
          "title": "query conditions"
        }
      }
    },
...
    "v1CreateSysUserReply": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "uint64"
        }
      }
    },
    "v1CreateSysUserRequest": {
      "type": "object",
      "properties": {
        "userName": {
          "type": "string",
          "title": "用户名昵称"
        },
        "userNumber": {
          "type": "string",
          "title": "用户编号"
        },
        "userPassword": {
          "type": "string",
          "title": "用户密码"
        },
        "userPhone": {
          "type": "string",
          "title": "用户手机号"
        },
        "userAvatar": {
          "type": "string",
          "title": "用户头像"
        },
        "createUser": {
          "type": "string",
          "title": "创建人"
        },
        "updateUser": {
          "type": "string",
          "title": "更新人"
        }
      }
    },
...
    "v1DeleteSysUserByIDReply": {
      "type": "object"
    },
    "v1DeleteSysUserByIDsReply": {
      "type": "object"
    },
...
    "v1GetSysUserByConditionReply": {
      "type": "object",
      "properties": {
        "sysUser": {
          "$ref": "#/definitions/v1SysUser"
        }
      }
    },
    "v1GetSysUserByConditionRequest": {
      "type": "object",
      "properties": {
        "conditions": {
          "$ref": "#/definitions/typesConditions"
        }
      }
    },
    "v1GetSysUserByIDReply": {
      "type": "object",
      "properties": {
        "sysUser": {
          "$ref": "#/definitions/v1SysUser"
        }
      }
    },
...
    "v1ListSysUserByIDsReply": {
      "type": "object",
      "properties": {
        "sysUsers": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1SysUser"
          }
        }
      }
    },
    "v1ListSysUserByLastIDReply": {
      "type": "object",
      "properties": {
        "sysUsers": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1SysUser"
          }
        }
      }
    },
    "v1ListSysUserByLastIDRequest": {
      "type": "object",
      "properties": {
        "lastID": {
          "type": "integer",
          "format": "uint64",
          "title": "last id"
        },
        "limit": {
          "type": "integer",
          "format": "int64",
          "title": "limit size per page"
        },
        "sort": {
          "type": "string",
          "description": "sort by column name of table, default is -id, the - sign indicates descending order."
        }
      }
    },
    "v1ListSysUserReply": {
      "type": "object",
      "properties": {
        "total": {
          "type": "integer",
          "format": "int64"
        },
        "sysUsers": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1SysUser"
          }
        }
      }
    },
    "v1ListSysUserRequest": {
      "type": "object",
      "properties": {
        "params": {
          "$ref": "#/definitions/typesParams"
        }
      }
    },
...
    "v1SysUser": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "uint64",
          "title": "用户ID"
        },
        "userName": {
          "type": "string",
          "title": "用户名昵称"
        },
        "userNumber": {
          "type": "string",
          "title": "用户编号"
        },
        "userPassword": {
          "type": "string",
          "title": "用户密码"
        },
        "userPhone": {
          "type": "string",
          "title": "用户手机号"
        },
        "userAvatar": {
          "type": "string",
          "title": "用户头像"
        },
        "createUser": {
          "type": "string",
          "title": "创建人"
        },
        "updateUser": {
          "type": "string",
          "title": "更新人"
        },
        "createdAt": {
          "type": "string",
          "title": "创建时间"
        },
        "updatedAt": {
          "type": "string",
          "title": "更新时间"
        }
      }
    },
...
    "v1UpdateSysUserByIDReply": {
      "type": "object"
    },
...
    "v1sysUserUpdateByIDBody": {
      "type": "object",
      "properties": {
        "userName": {
          "type": "string",
          "title": "用户名昵称"
        },
        "userNumber": {
          "type": "string",
          "title": "用户编号"
        },
        "userPassword": {
          "type": "string",
          "title": "用户密码"
        },
        "userPhone": {
          "type": "string",
          "title": "用户手机号"
        },
        "userAvatar": {
          "type": "string",
          "title": "用户头像"
        },
        "createUser": {
          "type": "string",
          "title": "创建人"
        },
        "updateUser": {
          "type": "string",
          "title": "更新人"
        }
      }
    }
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "description": "Type Bearer your-jwt-token to Value",
      "name": "Authorization",
      "in": "header"
    }
  }
}